fetchCookies is common for both users and post so separating it
